+++ __docs__/wonder-blocks-badge/badge-testing-snapshots.stories.tsx
@@ -9,6 +9,7 @@ import {
     StatusBadge,
     StreakBadge,
     DueBadge,
+    NeutralBadge,
 } from "@khanacademy/wonder-blocks-badge";
 import {IconMappings} from "../wonder-blocks-icon/phosphor-icon.argtypes";
 import {Icon, PhosphorIcon} from "@khanacademy/wonder-blocks-icon";
@@ -24,7 +25,7 @@ import {
 } from "../components/text-for-testing";
 import {View} from "@khanacademy/wonder-blocks-core";
 import {font, semanticColor, sizing} from "@khanacademy/wonder-blocks-tokens";
-import {Heading, HeadingLarge} from "@khanacademy/wonder-blocks-typography";
+import {Heading} from "@khanacademy/wonder-blocks-typography";
 import singleColoredIcon from "../components/single-colored-icon.svg";
 import {multiColoredIcon} from "../components/icons-for-testing";
 import Tooltip from "@khanacademy/wonder-blocks-tooltip";
@@ -131,15 +132,40 @@ export const StateSheetStory: StoryComponentType = {
 
         return (
             <View style={{gap: sizing.size_080}}>
-                <HeadingLarge>Badge</HeadingLarge>
+                <Heading size="large">Badge</Heading>
                 <StateSheet rows={rows} columns={columns} states={states}>
-                    {({props}) => <Badge {...props} />}
+                    {({props}) => (
+                        <View
+                            style={{gap: sizing.size_100, flexDirection: "row"}}
+                        >
+                            <Badge {...props} />
+                            <Badge {...props} showBorder={false} />
+                        </View>
+                    )}
                 </StateSheet>
-                <HeadingLarge>Status Badge</HeadingLarge>
+                <Heading size="large">Status Badge</Heading>
                 <StateSheet rows={statusRows} columns={columns} states={states}>
-                    {({props}) => <StatusBadge {...props} />}
+                    {({props}) => (
+                        <View
+                            style={{gap: sizing.size_100, flexDirection: "row"}}
+                        >
+                            <StatusBadge {...props} />
+                            <StatusBadge {...props} showBorder={false} />
+                        </View>
+                    )}
                 </StateSheet>
-                <HeadingLarge>Gem Badge</HeadingLarge>
+                <Heading size="large">Neutral Badge</Heading>
+                <StateSheet rows={rows} columns={columns} states={states}>
+                    {({props}) => (
+                        <View
+                            style={{gap: sizing.size_100, flexDirection: "row"}}
+                        >
+                            <NeutralBadge {...props} />
+                            <NeutralBadge {...props} showBorder={false} />
+                        </View>
+                    )}
+                </StateSheet>
+                <Heading size="large">Gem Badge</Heading>
                 <StateSheet
                     rows={rows}
                     columns={columnsWithShowIconProp}
@@ -147,7 +173,7 @@ export const StateSheetStory: StoryComponentType = {
                 >
                     {({props}) => <GemBadge {...props} />}
                 </StateSheet>
-                <HeadingLarge>Streak Badge</HeadingLarge>
+                <Heading size="large">Streak Badge</Heading>
                 <StateSheet
                     rows={rows}
                     columns={columnsWithShowIconProp}
@@ -155,7 +181,7 @@ export const StateSheetStory: StoryComponentType = {
                 >
                     {({props}) => <StreakBadge {...props} />}
                 </StateSheet>
-                <HeadingLarge>Due Badge</HeadingLarge>
+                <Heading size="large">Due Badge</Heading>
                 <StateSheet
                     rows={dueBadgeRows}
                     columns={columnsWithShowIconProp}
@@ -301,6 +327,7 @@ export const AllBadgesScenarios: StoryComponentType = {
             ...statusKinds.map((kind) => (
                 <StatusBadge label="Badge" kind={kind} />
             )),
+            <NeutralBadge label="Badge" />,
         ].map((component) =>
             React.cloneElement(component, {
                 icon: <PhosphorIcon icon={IconMappings.cookieBold} />,