+++ packages/wonder-blocks-card/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @khanacademy/wonder-blocks-card
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies [272bf31]
+    - @khanacademy/wonder-blocks-tokens@14.1.0
+    - @khanacademy/wonder-blocks-icon-button@10.5.5
+
+## 1.2.2
+
+### Patch Changes
+
+- 6d5c485: Include provenance information when publishing to npmjs
+- Updated dependencies [6d5c485]
+    - @khanacademy/wonder-blocks-core@12.4.1
+    - @khanacademy/wonder-blocks-icon-button@10.5.4
+    - @khanacademy/wonder-blocks-tokens@14.0.1
+
 ## 1.2.1
 
 ### Patch Changes