+++ packages/wonder-blocks-styles/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @khanacademy/wonder-blocks-styles
 
+## 0.2.34
+
+### Patch Changes
+
+- Updated dependencies [272bf31]
+    - @khanacademy/wonder-blocks-tokens@14.1.0
+
+## 0.2.33
+
+### Patch Changes
+
+- 6d5c485: Include provenance information when publishing to npmjs
+- Updated dependencies [6d5c485]
+    - @khanacademy/wonder-blocks-tokens@14.0.1
+
 ## 0.2.32
 
 ### Patch Changes