+++ packages/wonder-blocks-switch/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @khanacademy/wonder-blocks-switch
 
+## 3.3.22
+
+### Patch Changes
+
+- Updated dependencies [272bf31]
+    - @khanacademy/wonder-blocks-tokens@14.1.0
+    - @khanacademy/wonder-blocks-icon@5.3.2
+    - @khanacademy/wonder-blocks-styles@0.2.34
+
+## 3.3.21
+
+### Patch Changes
+
+- 6d5c485: Include provenance information when publishing to npmjs
+- Updated dependencies [6d5c485]
+    - @khanacademy/wonder-blocks-core@12.4.1
+    - @khanacademy/wonder-blocks-icon@5.3.1
+    - @khanacademy/wonder-blocks-styles@0.2.33
+    - @khanacademy/wonder-blocks-theming@4.0.1
+    - @khanacademy/wonder-blocks-tokens@14.0.1
+
 ## 3.3.20
 
 ### Patch Changes