+++ packages/wonder-blocks-tokens/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-tokens
 
+## 14.1.0
+
+### Minor Changes
+
+- 272bf31: Add feedback neutral semantic color tokens
+
+## 14.0.1
+
+### Patch Changes
+
+- 6d5c485: Include provenance information when publishing to npmjs
+- Updated dependencies [6d5c485]
+    - @khanacademy/wonder-blocks-theming@4.0.1
+
 ## 14.0.0
 
 ### Major Changes