+++ packages/wonder-blocks-toolbar/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @khanacademy/wonder-blocks-toolbar
 
+## 5.1.41
+
+### Patch Changes
+
+- Updated dependencies [272bf31]
+    - @khanacademy/wonder-blocks-tokens@14.1.0
+    - @khanacademy/wonder-blocks-typography@4.2.24
+
+## 5.1.40
+
+### Patch Changes
+
+- 6d5c485: Include provenance information when publishing to npmjs
+- Updated dependencies [6d5c485]
+    - @khanacademy/wonder-blocks-core@12.4.1
+    - @khanacademy/wonder-blocks-tokens@14.0.1
+    - @khanacademy/wonder-blocks-typography@4.2.23
+
 ## 5.1.39
 
 ### Patch Changes